requirements.txt
clevercsv==0.8.2
Django==4.2.16
dj-database-url==2.3.0
django-filter==24.3
django-modern-rpc==1.0.3
django-property-filter==1.2.0
djangorestframework==3.15.2
django-fernet-encrypted-fields==0.1.3
drf-access-policy==1.5.0
frozendict==2.4.6
gunicorn==23.0.0
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2-binary==2.9.10
requests==2.32.3
SQLAlchemy==1.4.54
whitenoise==6.7.0
